Silence clang 10 conversion warning
 src/opusfile.c:3242:18: warning: implicit conversion from 'unsigned
       int' to 'float' changes value from 4294967295 to 4294967296
       [-Wimplicit-int-float-conversion]
           r=seed*OP_PRNG_GAIN;
                  ^~~~~~~~~~~~
 src/opusfile.c:3179:29: note: expanded from macro 'OP_PRNG_GAIN'
 # define OP_PRNG_GAIN (1.0F/0xFFFFFFFF)
                            ~^~~~~~~~~~

+++ src/opusfile.c
@@ -3176,7 +3176,7 @@ static opus_uint32 op_rand(opus_uint32 _seed){
 
 # define OP_GAIN (32753.0F)
 
-# define OP_PRNG_GAIN (1.0F/0xFFFFFFFF)
+# define OP_PRNG_GAIN (1.0F/(float)0xFFFFFFFF)
 
 /*48 kHz noise shaping filter, sd=2.34.*/
